fix matrix jog left/right
NOTE: I'm assuming using notation timings for the notation version was correct,
even though the opposite was originally the case.  I didn't try to work out the
logic of what anything does, just what it says it's supposed to do.  I added a
bool just to make it clear what this parameter means, and I figure if it's
really wired backwards, we should flip the internal wiring so "use notation
timings" is the correct answer for what to do from notation.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11406 cf6a83e7-f017-4305-b0b3-678bde24f230