tweak the new strings slightly
One thing I find amusing about this one is that the English sounds a little off
due to how the translations come together, but Spanish sounds fine. "Bass down two
octaves clef" sounds like Franglais or Spanglish, but with Spanish word order it
comes out "clef of bass down two octaves" and that reads a lot better.

There's no practical solution to this one, and it's not a big deal.  I just
think it's funny.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11413 cf6a83e7-f017-4305-b0b3-678bde24f230