New "Cycle Slashes" feature in notation.
Feature freeze?  What feature freeze?  This is something I had quite forgotten I
intended to do, and when the inspiration struck me, I had a pleasant little time
throwing this together.  For once, something went smoothly from start to end,
and it all compiled and worked as expected the first time I tried.

I thought Classic had a missing / shortcut, and I was intending to hook it back
up and add a companion to my new "cycle dots" icon in the right-side toolbar
group.  It turns out Classic never had this, and so it's all a totally new
feature that I think we should have done ages ago.

It can be used in a variety of ways.  One method is to use it as a postfix
compliment to one of the do/I re/II mi/III keys, so you can enter, A/ S/// D//
to get notes with 1, 3 and 2 slashes.  Another method is to select a group of
notes and click the icon or hit / the desired number of times.  Finally, the
more dubious mode of operation is to select a group of notes that have random
numbers of slashes, and cycle through 0-5 on a note per note basis.  I don't
think that's particularly useful, but having this new command work event by
event and just cycle from 0-5 one by one as it iterated along seemed like a
joyfully simple way to go.

(I also worked over the "cycle dots" icon to make it look less clunky.  These
are still faintly clunky, but I think they're better.)


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11448 cf6a83e7-f017-4305-b0b3-678bde24f230