It was probably always wrong to ignore segment transposition in the matrix, but it probably never became an issue because people working purely with MIDI tend to use the matrix exclusively, and these people tend to view segment transposition as more of a sound effect than a fundamental property.
Now that we have per-key pitch highlights in the matrix, and now that we have
the possibility of displaying many segments that might each have a different
transposition on the same matrix grid, it seems like the only sane way to to
handle these various considerations is to make the matrix a concert pitch
display that represents the notes on the piano keyboard the same way a piano
player would see them when faced with each of these transposed parts.

If an Eb instrument is playing a G major scale in the key of G, the piano sees
this for the Bb major scale it really is.  For an Eb segment (-9 or +3) with a
key change to G major, the pitch highlights should be for Bb major, and the
notes of the scale should line up with those highlights.

This should be meshed with the way we handle transposition in MIDI performance,
which is that we record the MIDI pitches you play, and then we adjust the
notation against the track-level "create segments with" transposition property
such that when it is reproduced later, you hear the same notes you played.  This
means that when entering notes with the pencil into a transposed segment, they
need to have their pitches adjusted appropriately.

All of this turns out to be quite difficult to do accurately, because events
know nothing of their owning segments, and the transpose factor is a property of
the segment.  This requires much juggling to get right, and I've reverted my
patch several times to go try some other approach, only to restore it. I think
this is the horse to bet on, but this work should be completed in a branch, so I
can better benefit from version control.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/concert-pitch-matrix@11456 cf6a83e7-f017-4305-b0b3-678bde24f230