The moving around bits are sorted now, and all initial design requirements are satisfied successfully.  I'm going to ramp the testing up a notch and merge back to trunk/
I'm breaking off to explore a bug I discovered, where you can click on matrix
elements that don't belong to the active segment, start dragging them, and in so
doing create new events in the current segment.

(All in all a very arduous, mind-bending experience, but a very gratifying one.
I think the concert pitch matrix is the ideal solution to these problems, and
not a hack compromise.  It's nice reaching for the tall fruit and plucking it!)


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/concert-pitch-matrix@11469 cf6a83e7-f017-4305-b0b3-678bde24f230