OK, this has got it.  You can no longer click elements outside of the current segment to create new notes out of thin air.  Read comments if you're interested in the details.
I think this is workable forcing users to pre-select which segment to
manipulate, and I think moreover it would be groovy to throw together some
convenience widget to supplement Alt+PgUp/Down when people are in more of a
clicky mood than a typing mood

It gives me an idea for a new active staff/segment navigation widget I'm going
to go ahead and play with here in this branch.  I'm thinking to put it in the
panner block, on the left, and use another wheel widget.  Both notation and
matrix really only need one control, for flipping active segments.  Active
staffs in notation are very easy to control at will, but segments not so much.

(I shouldn't get distracted with a new nugget, and should instead spend tomorrow
either fixing bugs (of which there are still plenty) or working on the
documentation, and I definitely shouldn't break the string and feature freeze
this close to releasing the beta, but watch me.  I'm doing it anyway.  Damn
protocol.  This idea is too good, and too easy to implement, to put off.
Although I *will* go to bed now and at least put it off for tomorrow.)


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/branches/concert-pitch-matrix@11471 cf6a83e7-f017-4305-b0b3-678bde24f230