Trying to add a configuration option for this lead to an interesting discovery. As a result, the previous patch has been backed out completely, and this new version uses a configuration option to optionally turn on the old, buggy behavior.
I really don't like this option now that I've studied all of this code, and
looked at how random applications that load files from random places deal with
this situation.  Nobody displays the entire URL in the title bar that I can
find.  It's an option to turn on a bug that was clearly a departure from an
original intent based on reasonable best practices.

It's a long-standing and very visible bug though, so I understand Chris
Cherrett's case for keeping it around.  This version of reworking everything
provides a config option on the general options Presentation tab that allows the
buggy behavior to be restored.

This means that because the bug was in (main window)::openURL(), the only time
the window title was set to the full path was in this one place.  Elsewhere, it
was always the intended (QFileInfo).fileName(), and you can see this behavior in
Rosegarden Classic.  Save a file "bleargy.rg" and the title becomes (gruesomely
enough) "Rosegarden: bleargy.rg - Rosegarden" rather than the full path.  The
full path only happens when loading something pre-existing via openURL().

I'm not much interested in making the "restore the old bug" case do a better job
of ensuring the full path is always visible when that option is turned on.
Let's leave it at turning the bug back on for the most dedicated old-time users,
and quit there.  This behavior was clearly always a goof in first place, because
the "long path" implementation was so broken.

So odd that it took almost 10 years before anyone noticed!


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11499 cf6a83e7-f017-4305-b0b3-678bde24f230