I really shouldn't commit a buggy new feature to trunk/ two days before releasing the beta, but I'm cramming it in anyway.  I want this to get out there, right into the middle of things, and if it works at all it's good enough to be worth looking at as a taste of what the final release will have.
Hot gee ga damn!

This one was fun.  I had a flash of inspiration while answering a FAQ for the 11
millionth time, and I finally settled on a way to do something I had been
chewing on somewhere in the back of my mind for years.

Using overlapping segments to contain different voices on the same staff has
always sucked because:

 - You have to go out to the canvas to draw the thing, figure out where to
   put it, and futz around getting it into an edit view along with the other
   segment you're working on

 - You have to do something with the color manually, or else the various tools
   that use color to indicate that the active segment has changed are not very
   useful

With this commit, I consider the solution workable and worth seeing through,
although there's no way it's going to get cleaned up and polished before Friday.

  - Hit the 1/2/3 add layer button
  
  - Roll the segment changer widget

  - Boogy down

The add layer button uses the currently active segment as a template for
creating an empty new one of the same duration, directly overlapping it, but in
a different color.  Then it "reboots" the notation view, so you can begin using
it immediately.

It actually works once.  Using it a second time causes a nasty crash I haven't
tracked down yet.  This fact makes me think it would be most prudent to start a
branch and work through this more off to the side, but after a successful day
like this, I'm feeling reckless, ambitious, and optimistic that I can pull this
off in the nick of time.  I have 48 hours and change, and nothing better to do
with this desolate holiday.

I'm going for it.  If I have to, I'll back up and fork off a stable branch as of
the revision before this one to release from, or something, but I'm probably one
showstopper away from the coolest thing I've ever thought of and then executed
in one day, and this little cherry on top is my present to myself.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11502 cf6a83e7-f017-4305-b0b3-678bde24f230