* cleaner code to update IPB
GUI update is fine, but must select tracks individually to push studio change
to track.

I've spend many hours trying to figure out how to push the sendBank,
sendProgram, etc. to the sequencer -- I think that is the problem.

Here is the sequence:
I load the sample brandenburg midi
I import my default studio
IPB updates correctly.
Press Play
the IPB says it should be sending Bank changes but it isn't
I select each of the tracks with the mouse.
When I do the program the program, and pank are sent and the voice
changes.

I select all of them individually and the same thing happens.

Some help with how to push all of the updated device instruments to the
sequencer or composition would be very helpful.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11514 cf6a83e7-f017-4305-b0b3-678bde24f230