Resurrect that patch for the graphics system changer, and iron out the kinks that had caused me to dump it originally.
I don't much like having an obscure option right in everyone's face at the top
of the first page, but a substantial proportion of user problems reported during
the alpha and beta testing phases have been curable by changing to a different
graphics system.  I'm afraid putting an easy control for this into people's
hands is a must.

I've tried to make it a bit less technical and jargony by referring to the
graphics systems as Normal, Safe, and Fast instead of Raster, Native and OpenGL,
putting refrences to what they really are in the tooltip for people more in the
know.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11517 cf6a83e7-f017-4305-b0b3-678bde24f230