Groundwork for subclassing QPlastiqueStyle as Rosegarden::ThornStyle in order to override a number of default style icons.
I started exploring this when I finally had it with how bad the (i) icon
looks in our message boxes, but this will allow me to twiddle a number of minor
details I couldn't touch with stylesheets (within or without), and it should be
worthwhile.

I'm quitting for now, but I've got the code in Qt mapped out I'm going to copy
and doctor up, and it should all be rather straightforward the next time I get a
chance to play.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11518 cf6a83e7-f017-4305-b0b3-678bde24f230