The other half of the experimental work in removing the channel controls.  If we find a document whose MIDI instruments were mapped to channels other than the defaults, remap the channels to enforce the new obligatory correspondence and throw an information dialog explaining what happened.
The more I think about this whole idea, the more right it seems.  You could have
two different instruments pointing to the same channel before, and that's just
wrong.  Two different reverb/volume/pan settings in two different instruments
transmitting on the same channel.  Which one takes precedence?  It's undoubtedly
responsible for at least one report of something not behaving correctly, and
this never should have been allowed in the first place.  I don't think we had a
clear idea of how the whole model should work in the beginning.  A channel
control makes sense when this information belongs directly to a track, but our
instrument concept demands something different, and I believe the current
implementation will fit the bill.

I don't expect this to be noticed by very many people, because most people used
the instrument to set the channel anyway.  We do need to audit all the docs and
clarify how all of this works now.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11528 cf6a83e7-f017-4305-b0b3-678bde24f230