One solid improvement.  Don't have combo boxes like:
      Device     [         General MIDI Device           ]
      Instrument [ General MIDI Device #1 (Acoustic Grand]

Instead, like:

      Device     [         General MIDI Device           ]
      Instrument [      #1 (Acoustic Grand Piano)        ]

That's only the beginning of a very ugly looking slog to iron out the problems
with this though.

It looks like the lists of names and device ids are out of sync, and the pattern
isn't very obvious at all.

The TPB ignores it when the instrument is reassigned via the track button menu.

Changes made in the TPB don't propagate to the track button menu properly.  For
instance, you pick "Audio" and the IPB shows that it's really Synth Plugin, and
the track button LED is orange, but the track button menu shows General MIDI as
the connected instrument.

It's a real headache.  I think it may very well be easier to fix this by writing
a new implementation.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11538 cf6a83e7-f017-4305-b0b3-678bde24f230