Regression point.
I think I'm finally zeroing in on it.  TB sends 2010 for MIDI #9 (in test
environment) and TPB searches for 2010 and finds it within its local
representation, but it finds 2010 for device __0__ and not 2000.

Looking at XML, this seems correct.  Device 0 really is the one that has
instrument 2000.

    <device id="0" name="General MIDI Device"
        ....
        <instrument id="2000" channel="0" type="midi">

This code gets device 1000 for audio and 10000 for synth from Device::getId()
which seems damn bizarre.  It should be device 14 for instrument 10000 in this
example.  It's getting the first instrument in the device from Device::getId()
which makes no sense at all, and shouldn't have changed since Classic.

That's the key anyway.  I think the key is in Device.  Some bizarre side effect
of the "don't create extra garbage" change.

Maybe not, but there's something like that afoot here.  Either fix Device or
else hack this to get the first instrument for device n as what it treats to be
a device ID to match what it's doing for audio and synth already.  One or the
other should work.

If it really is in Device, then why do the TB controls still work?  Maybe
because they *don't* care about the device, and just ultimately lead you to
choose instrument 1 to n from an assembled list of all available.  That's why
they emit an instrument number to the TPB.  But maybe it's not that simple
either.

Bleargh.  What a confusing nightmare to fix two stupid combo boxes I would just
as soon disable and remove as go through all this.  Users.



git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11544 cf6a83e7-f017-4305-b0b3-678bde24f230