Misery.  I was so close I could almost taste it, and then I lost the thread and got so horribly befuddled I just had to give up.
Device really is returning 0, 10000, 1000 for the device IDs.

The track buttons have a flat list of instruments from 0 to n, where 0 is "Audio
#1" and n is "MIDI #16"

We need to figure out which index to send that flat list to get it to choose the
instrument that corresponds with what we've dialed into our controls here.  The
code used to work by iterating through the available devices, and adding the x
instruments each device contained to the base index.  This seems to have assumed
an ordering that is no longer true.  Audio and MIDI are flipped, but it's
something more than that.  It's some assumption that just isn't true, but quite
what keeps escaping my grasp.

Having poured 20 hours down this rabbit hole now, I'm pretty close to letting
this one win.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11545 cf6a83e7-f017-4305-b0b3-678bde24f230