Make the "Safe" native graphics system the default configuration option.
I've just got yet another tester reporting crashes in Qt code over which we have
no direct control, and unlike some of the other problems that were a little more
mysterious, this one was crystal clear where it was pointing the finger:

#3  0xb6e217e5 in QRasterWindowSurface::~QRasterWindowSurface() () from
  /usr/lib/libQtGui.so.4 #4  0xb6e3eb35 in ?? () from /usr/lib/libQtGui.so.4

This means about half the current testers have had problems with the raster
graphics system, and I just don't think it's reliable enough to be the default.
I've been using the raster system myself for going on a year now with no major
issues since we cleaned up the segment canvas scrubbing problem, but my
experience is only shared by about half of the current sample.

I'll add a FAQ entry tomorrow about bad graphics performance, and what to try to
see if it works.  Maybe this will settle down in the future, and we can make
raster the default again, or even opengl, but apparently that time is not yet.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11557 cf6a83e7-f017-4305-b0b3-678bde24f230