The new ProgressDialog does something when pressed forcefully now, although real ones still just flicker black briefly and disappear (and half the code is still commented out too.)
Used ProgressBar throughout, and experimented with style a bit.

Finally remembered to fix InputDialog's style so it's more sensible.  I'm not
yet convinced about this whole business of using "Rosegarden" as the window
title throughout, and taking what used to be the old title and putting it in an
<h3> block above the usual guts of the dialog.  It has some potential, but it
all seems a little off still.  I'm not so much worried about using "Rosegarden"
as the title because that's what Windows does, as because we have umpty dialogs
with titles like "[o]Select a p[_][*][x]" anyway.  These titles are largely
pointless, and in many cases, if the text is worth displaying at all, we need to
do something more sensible with it.  Maybe all I need is to try <h4> or so and
make the texts a bit smaller, but we've used <h3> in other contexts that have
been translated, and I'm just going to go with this like it is for now, and not
get sucked into nitpicking minutiae to death in the 11th hour.

Although, reminder to self: I want to do another warning widget icon about being
in the crappy graphics mode (wow, how '80s that sounded), and suggest users try
their luck with a better one.  This one probably shouldn't even trigger the /!\
warning, although I'm not sure how anyone would ever notice it if it didn't.
Hrm.  Weighing that, it's probably still just too obnoxious to do anything more
than simply have some kind of little status icon.

Heee...  A little picture of a "safe" for safe mode.  That will go great with
our little picture of a llave that means tecla.  Maybe I should just use the cow
icons here.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11567 cf6a83e7-f017-4305-b0b3-678bde24f230