Waypoint.  Never manipulate the progress dialog's progress bar directly.  Go through the dialog.
No real improvement in function yet, but I'm onto something promising.  Sensible
looking stuff is getting generated and emitted, and I've got to re-route it to
where it was trying to get, right after this cleanup commit.

NOTE: I also rediscovered CountdownDialog in the midst of this.  I decided not
to excise it because it's in deeply enough that it would take more thought than
I have to spare at the moment to work out how to remove every trace of it.  What
bits are only needed for that dialog, and nothing else, etc.?

A good afternoon project for someone after Thorn and/or after we fork a stable
release branch (not just yet) to piddle with.  Root this thing out and throw it
away.  (For those who don't know, countdown dialogs told you how much longer you
had before you face planted into the brick wall at the end of the composition.
Usually right in the middle of a good improv solo that didn't get recorded,
because you were jamming and not paying attention to where the brick wall was.)


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11599 cf6a83e7-f017-4305-b0b3-678bde24f230