A large measure of sorting on the whole progress bar issue.  The heart of all of this was that RosegardenMainWindow had a progressBar() and m_progressBar() that looks to have been there and called that for historical reasons.  It looks like "building a KDE app for the first time" code to me.
I can't quite tell if this m_progressBar was ever meant to show any kind of
progress in addition to being a CPU meter.  Memory says it probably was, and
changed modes for certain purposes.  I think there are still vestiges of that
use left.  Anyway, forget trying to sort out what was.  What happened was
somebody replaced a vast number of (KProgressDialog*)->getProgressBar()->foo()
with m_progressBar foo, confusing the progress bar inside the progress dialog
that we used to access from outside for some reason with the progress bar being
used (at least primarily) as the CPU meter.  The result was all this progress
was being reported under our noses the whole time, and it was going to the wrong
bar.

I've sorted that out, and we're going to use ProgressDialog for reporting
progress, and m_cpuBar for the CPU meter.  If there's some crossed purpose that
was intentional that I've missed, we'll build it some other way, such as by
chaining the progress bar in the progress dialog to the CPU/progress meter.  (If
that was intentional and is desired.  Otherwise I'm not going to fool with it.
If anyone knows or has an opinion, please speak up.)


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11602 cf6a83e7-f017-4305-b0b3-678bde24f230