Fix #2929065: re-parent the offending message box to 0 instead of this.  It now inherits the correct style, and there seems to be no real consequence.
I also changed the parents of several other message boxes in the same vicinity,
and consolidated all versions of the "audio not writable" to use
QMessageBox::Yes.  I think the unusual "Yes/Cancel" may be just the thing to
give a little jolt and make someone take time to actually read the message box.


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11613 cf6a83e7-f017-4305-b0b3-678bde24f230