Finish this out with some tmp file open/write test whizrums.
I'm in a hurry, but it passes owned by different user, chmod -x, chmod -w, chmod
-r.  Considering the most important thing in this was simply creating the
directory (to avoid problems with things failing due to it not existing) this is
probably robust enough to be pretty solid treatment.

Note that I thought about some "does not exist, would you like to create?"
business instead, but that would have required an irritating popup.  Using
Classic on a box with bad everything, I'm really starting to appreciate the
lengths I've gone to with Thorn to avoid these damn things.  Much cleaner!

Well, off to the rat race.  :(


git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk/rosegarden@11635 cf6a83e7-f017-4305-b0b3-678bde24f230